stop using #ord and create ::cipher_helper
The problem with using #ord is that I found it
difficult to only include valid characters. Cases
would return invalid chars, and then they couldn't
be deciphered back. I instead made characters
shift with an hard-coded array. This makes it easy
to include characters that we'd like to shift.

I noticed that ::cipher and ::decipher only had
one different character. To adhere to DRY, I kept
one copy as a helper method with an extra variable
for the difference. I'm not sure how maintainable
this is. I'll learn a better method eventually.